backend/requirements-test.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements-test.in
#
attrs==22.1.0
    # via pytest
iniconfig==1.1.1
    # via pytest
packaging==21.3
    # via pytest
pluggy==1.0.0
    # via pytest
py==1.11.0
    # via pytest
pyparsing==3.0.9
    # via packaging
pytest==7.1.3
    # via
    #   -r requirements-test.in
    #   pytest-django
pytest-django==4.5.2
    # via -r requirements-test.in
six==1.16.0
    # via -r requirements-test.in
tomli==2.0.1
    # via pytest
